requirements.txt
fastapi
uvicorn
python-dotenv
crewai
masumi
pydantic
python-multipart
httpx
langchain_openai
pysqlite3-binary
sentence-transformers
langchain-google-genai